+++ debug-app/src/main/res/values/constants.xml
@@ -1,5 +1,5 @@
 <resources>
     <string name="rover_endpoint">https://api.rover.io</string>
     <!-- Put your own Rover API token here! -->
-    <string name="rover_api_token">BLANK</string>
+    <string name="rover_api_token">72daccff7a7d497de0de3a8d3757dba99f8131e1</string>
 </resources>
