+++ experiences/src/main/kotlin/io/rover/experiences/ui/navigation/ExperienceNavigationViewModel.kt
@@ -2,7 +2,6 @@ package io.rover.experiences.ui.navigation
 
 import android.arch.lifecycle.Lifecycle
 import android.arch.lifecycle.LifecycleObserver
-import android.arch.lifecycle.LifecycleOwner
 import android.arch.lifecycle.OnLifecycleEvent
 import android.os.Parcelable
 import io.rover.core.data.domain.AttributeValue
@@ -20,7 +19,6 @@ import io.rover.core.streams.flatMap
 import io.rover.core.streams.map
 import io.rover.core.streams.shareHotAndReplay
 import io.rover.core.streams.subscribe
-import io.rover.core.streams.takeUntil
 import io.rover.core.tracking.SessionTrackerInterface
 import io.rover.experiences.data.domain.events.asAttributeValue
 import io.rover.experiences.ui.containers.ExperienceActivity
@@ -223,13 +221,10 @@ open class ExperienceNavigationViewModel(
             )
         }
 
-        // session changes from moving between screens or leaving
+        // session changes from moving between screens.
         screenSubject.subscribe { screenUpdate ->
             trackLeaveScreen()
             trackEnterScreen(screenUpdate.screenViewModel)
-        }
-        externalNavigationEvents.subscribe { _ ->
-//            trackLeaveScreen()
         }
 
         // handle visibility changes for session tracking:
@@ -237,17 +232,22 @@ open class ExperienceNavigationViewModel(
         activityLifecycle.addObserver(object : LifecycleObserver {
             @OnLifecycleEvent(Lifecycle.Event.ON_RESUME)
             fun presented() {
+                trackEnterExperience(experience)
+
+                // if an experience screen is already active (which can happen if the containing
+                // Activity is resumed without having been entirely destroyed), then handle
+                // re-emitting the trackEnterScreen event.
                 activeScreenViewModelIfPresent().whenNotNull { activeScreen ->
                     trackEnterScreen(
                         activeScreen
                     )
                 }
-
             }
 
             @OnLifecycleEvent(Lifecycle.Event.ON_PAUSE)
             fun dismissed() {
                 trackLeaveScreen()
+                trackLeaveExperience(experience)
             }
         })
     }
@@ -303,6 +303,23 @@ open class ExperienceNavigationViewModel(
         } ?: closeExperience(state.backStack) // can't go any further back: backstack would be empty; instead emit Exit.
     }
 
+    protected fun trackEnterExperience(experience: Experience) {
+        sessionTracker.enterSession(
+            ExperienceSessionKey(experience.id.rawValue, experience.campaignId),
+            "Experience Presented",
+            "Experience Viewed",
+            sessionExperienceEventAttributes(experience)
+        )
+    }
+
+    protected fun trackLeaveExperience(experience: Experience) {
+        sessionTracker.leaveSession(
+            ExperienceSessionKey(experience.id.rawValue, experience.campaignId),
+            "Experience Dismissed",
+            sessionExperienceEventAttributes(experience)
+        )
+    }
+
     /**
      * Called when leaving an Experience screen.  Add any side-effects here, such as tracking the
      * session.
@@ -316,7 +333,7 @@ open class ExperienceNavigationViewModel(
             sessionTracker.leaveSession(
                 ExperienceScreenSessionKey(experience.id.rawValue, currentScreenId),
                 "Screen Dismissed",
-                sessionEventAttributes(screenViewModel)
+                sessionScreenEventAttributes(screenViewModel)
             )
         }
     }
@@ -330,7 +347,7 @@ open class ExperienceNavigationViewModel(
             ExperienceScreenSessionKey(experience.id.rawValue, screenViewModel.screenId),
             "Screen Presented",
             "Screen Viewed",
-            sessionEventAttributes(screenViewModel)
+            sessionScreenEventAttributes(screenViewModel)
         )
     }
 
@@ -378,7 +395,13 @@ open class ExperienceNavigationViewModel(
         )
     }
 
-    private fun sessionEventAttributes(screenViewModel: ScreenViewModelInterface): Attributes {
+    protected open fun sessionExperienceEventAttributes(experience: Experience): Attributes {
+        return hashMapOf(
+            Pair("experience", experience.asAttributeValue())
+        )
+    }
+
+    protected open fun sessionScreenEventAttributes(screenViewModel: ScreenViewModelInterface): Attributes {
         return hashMapOf(
             Pair("experience", experience.asAttributeValue()),
             Pair("screen", screenViewModel.attributes)
@@ -396,10 +419,19 @@ open class ExperienceNavigationViewModel(
     ) : Parcelable
 
     /**
-     * A unique descriptor for identifying a given experience screen to [SessionTrackerInterface].
+     * A unique descriptor for identifying a given experience screen to the [SessionTrackerInterface].
      */
     data class ExperienceScreenSessionKey(
         private val experienceId: String,
         private val screenId: String
     )
+
+    /**
+     * A unique descriptor for identifying a given experience to the [SessionTrackerInterface].
+     */
+    data class ExperienceSessionKey(
+        val experienceId: String,
+        val campaignId: String?
+    )
+
 }
